requirements.txt
glfw==2.8.0
imgui==2.0.0
numpy==1.24.4
pillow==10.4.0
pygame==2.6.1
PyOpenAL==0.7.11a1
PyOpenGL==3.1.9
PyWavefront==1.3.3
